Adding dtos

+++ src/main/java/com/ts/projekt_ts/infrastucture/service/BookService.java
@@ -1,11 +1,15 @@
 package com.ts.projekt_ts.infrastucture.service;
 
+import com.ts.projekt_ts.controllers.dto.CreateBookDto;
+import com.ts.projekt_ts.controllers.dto.CreateResponseDto;
+import com.ts.projekt_ts.controllers.dto.GetBookDto;
 import com.ts.projekt_ts.infrastucture.entity.BookEntity;
 import com.ts.projekt_ts.infrastucture.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class BookService {
@@ -15,17 +19,30 @@ public class BookService {
         this.bookRepository = bookRepository;
     }
 
-    public List<BookEntity> getAll(){
-        return bookRepository.findAll();
+    public List<GetBookDto> getAll(){
+        var books =  bookRepository.findAll();
+        return books.stream().map((book) -> new GetBookDto(book.getId(), book.getIsbn(), book.getTitle(), book.getAuthor(), book.getPublisher(), book.getPublicationYear(), book.getAvaliableCopies() > 0)).collect(Collectors.toList());
     }
-    public BookEntity getOne(long id){
-        return bookRepository.findById(id).orElseThrow(() -> new RuntimeException("Book not found"));
+    public GetBookDto getOne(long id){
+        var book = bookRepository.findById(id).orElseThrow(() -> new RuntimeException("Book not found"));
+        return new GetBookDto(book.getId(), book.getIsbn(), book.getTitle(), book.getAuthor(), book.getPublisher(), book.getPublicationYear(), book.getAvaliableCopies() > 0);
     }
 
-    public BookEntity create(BookEntity book){
-        return bookRepository.save(book);
+    public CreateResponseDto create(CreateBookDto book){
+        var bookEntity = new BookEntity();
+        bookEntity.setAuthor(book.getAuthor());
+        bookEntity.setIsbn(book.getIsbn());
+        bookEntity.setTitle(book.getTitle());
+        bookEntity.setPublicationYear(String.valueOf(book.getPublicationYear()));
+        bookEntity.setPublisher(book.getPublisher());
+        bookEntity.setAvaliableCopies(String.valueOf(book.getAvaliableCopies()));
+        var newBook = bookRepository.save(bookEntity);
+        return new CreateResponseDto(newBook.getId(), newBook.getIsbn(), newBook.getTitle(), newBook.getAuthor(), newBook.getPublisher(), newBook.getPublicationYear(), newBook.getAvaliableCopies());
     }
     public void delete(long id){
+        if(!bookRepository.existsById(id)){
+            throw new RuntimeException("Book not found");
+        }
         bookRepository.deleteById(id);
     }
